Add ability to merge a story.
Clicking on a story in the list of unmerged stories now pops up a dialog asking
the user to confirm they want to merge the story. If they do, the story will be
merged. The list of pending stories will automatically be refreshed.

This is still a work in progress. Only the list of unmerged stories is refreshed,
the list of merged stories is not automatically updated yet. The unmerged/merged
lists share the StoryListFragment. Therefore the same dialog is popping up when
a merged story is clicked. The dialog does not yet identify which list is displayed
and still offers the user the chance to merge the already merged story. The dialog
needs to be extended to deal with both scenarios.